Simplify the test case somewhat
I have narrowed down the mapping attributes by removing
those which don't seem to affect the outcome between NH 5.4.2 and 5.4.3.
The only one which seems to make a difference is not-found="ignore".
It's also not specific to using 'formula' mappings, it occurs
when using a normal named column as well.